// src/interfaces/common.ts
export enum ETheme {
  DEFAULT = "DEFAULT",
  LINE = "LINE",
  DANGER = "DANGER",
}

export enum ESize {
  LARGE = "LARGE",
  MEDIUM = "MEDIUM",
  SMALL = "SMALL",
}

export enum EThemeTag {
  HTML = "HTML",
  CSS = "CSS",
  SCSS = "SCSS",
  JQUERY = "JQUERY",
  JAVASCRIPT = "JAVASCRIPT",
  REACT = "REACT",
  VUE = "VUE",
  BOOTSTRAP = "BOOTSTRAP",
}